+++ backend/src/main/java/com/interview/common/error/ApiExceptionHandler.java
@@ -27,7 +27,7 @@ public class ApiExceptionHandler {
                 .stream()
                 .collect(Collectors.toMap(
                         FieldError::getField,
-                        FieldError::getDefaultMessage,
+                        fieldError -> fieldError.getDefaultMessage() != null ? fieldError.getDefaultMessage() : "Invalid value",
                         (left, right) -> left
                 ));
