+++ backend/src/test/java/com/interview/workorder/mapping/WorkOrderMapperUnitTest.java
@@ -5,8 +5,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import com.interview.customer.entity.Customer;
 import com.interview.workorder.entity.WorkOrder;
 import com.interview.workorder.model.WorkOrderStatus;
-import com.interview.workorder.request.WorkOrderRequest;
-import com.interview.workorder.response.WorkOrderResponse;
+import com.interview.workorder.dto.WorkOrderRequest;
+import com.interview.workorder.dto.WorkOrderResponse;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.Test;
 import org.mapstruct.factory.Mappers;
