+++ backend/src/test/java/com/interview/workorder/service/WorkOrderServiceUnitTest.java
@@ -14,8 +14,8 @@ import com.interview.workorder.dao.WorkOrderRepository;
 import com.interview.workorder.entity.WorkOrder;
 import com.interview.workorder.mapping.WorkOrderMapper;
 import com.interview.workorder.model.WorkOrderStatus;
-import com.interview.workorder.request.WorkOrderRequest;
-import com.interview.workorder.response.WorkOrderResponse;
+import com.interview.workorder.dto.WorkOrderRequest;
+import com.interview.workorder.dto.WorkOrderResponse;
 import java.util.List;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
